Addresses HIGH/MEDIUM findings from the phase 2 deep review: IDOR scope
guard on submit_answer, UUID coercion and input validation, PII removal
from rate-limit log, SGD chain rule correctness in scheduler, atomic
counter and bulk updates in recalibrator (plus race-safe ON CONFLICT
bootstrap), DB expression index + FK ondelete + type alignment + missing
indexes, frontend a11y/type safety/mutation error handling on the
recalibration UI.

Conflict resolution:
- rate_limit.py: kept HEAD's PII-free log format (semantically equivalent)
- worker.py: kept HEAD entirely — phase1a+1b's session-isolated worker
  with stuck-task reset already supersedes phase2's pre-1b version.
- Rebased 3c7d5e9f0a11_phase2_review_index_and_fk_fixes.py down_revision
  onto a1b2c3d4e5f6 (phase1a) to keep the migration chain linear.

+++ backend/app/models/flashcard.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, Boolean, DateTime, Float, ForeignKey, Integer, Numeric, String, UniqueConstraint, func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -81,8 +81,9 @@ class FlashcardProgress(UUIDPrimaryKeyMixin, Base):
     repetitions: Mapped[int] = mapped_column(Integer, default=0)
     next_review: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     last_reviewed: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
-    # FSRS-5 state columns (null until first FSRS review)
-    stability: Mapped[float | None] = mapped_column(Float, nullable=True)
-    difficulty: Mapped[float | None] = mapped_column(Float, nullable=True)
+    # FSRS-5 state columns (null until first FSRS review). Stored as Numeric to
+    # match the underlying DDL exactly and avoid spurious autogenerate diffs.
+    stability: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    difficulty: Mapped[Decimal | None] = mapped_column(Numeric(4, 2), nullable=True)
     last_grade: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    fsrs_review_count: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
+    fsrs_review_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
